requirements.txt.txt
python-telegram-bot==20.7
fastapi
uvicorn
ebooklib
beautifulsoup4
pymongo
